Handle all values from a rowid database column as long (bgo#696520)
Update to the latest Hyena from git master, to bring in the underlying
fix for bgo#696520.

As a consequence, all values coming from a database column defined as
'INTEGER PRIMARY KEY' should now be stored in variables typed as 'long".

This is overkill for some cases, like PrimarySourceId, and brings some
slight increase in memory usage. But I think the benefits are worth it:
consistency in the codebase, no Convert.ToInt32 calls here and there,
and no strange bugs like bgo#696520.

+++ src/Core/Banshee.Services/Banshee.Collection.Database/Bookmark.cs
@@ -55,7 +55,7 @@ namespace Banshee.Collection.Database
         public long BookmarkId { get; private set; }
 
         [DatabaseColumn]
-        protected int TrackId {
+        protected long TrackId {
             get { return Track.TrackId; }
             set {
                 Track = DatabaseTrackInfo.Provider.FetchSingle (value);
