+++ src/Core/Banshee.Services/Banshee.Collection.Database/DatabaseArtistInfo.cs
@@ -151,8 +151,8 @@ namespace Banshee.Collection.Database
         }
 
         [DatabaseColumn("ArtistID", Constraints = DatabaseColumnConstraints.PrimaryKey)]
-        private int dbid;
-        public int DbId {
+        private long dbid;
+        public long DbId {
             get { return dbid; }
         }
